scripts/requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Security and rate limiting
Flask-Limiter==3.5.0
Flask-CORS==4.0.0

# API documentation
flask-swagger-ui==4.11.1
flasgger==0.9.7.1